[1630]: move fixtures to conftest
